Converted Layout to a component
