+++ src/components/layout.js
@@ -1,0 +1,55 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import Helmet from 'react-helmet'
+import { StaticQuery, graphql } from 'gatsby'
+
+import Header from './header'
+import './layout.css'
+
+const Layout = ({ children, data }) => (
+    <StaticQuery
+        query={graphql`
+            query SiteTitleQuery {
+                site {
+                    siteMetadata {
+                        title
+                    }
+                }
+            }
+        `}
+      render={data =>(
+        <>
+            <div>
+                <Helmet
+                    title={data.site.siteMetadata.title}
+                    meta={[
+                        { name: 'description', content: 'Sample' },
+                        { name: 'keywords', content: 'sample, something' },
+                    ]}
+                >
+                    <html lang="en"/>
+                </Helmet>
+                <Header siteTitle={data.site.siteMetadata.title} />
+                <div
+                style={{
+                    margin: '0 auto',
+                    maxWidth: 960,
+                    padding: '0px 1.0875rem 1.45rem',
+                    paddingTop: 0,
+                }}
+                >
+                    {children}
+                </div>
+            </div>
+        </>
+      )}
+    />
+)
+
+Layout.propTypes = {
+  children: PropTypes.node.isRequired,
+}
+
+export default Layout
+
+
